Create Math Service
